complete authentication and authorization using jwt token generated from logged in user

+++ controllers/api/v1/posts_api.js
@@ -30,36 +30,27 @@ module.exports.index=async function(request,response){
 }
 
 
-
+ 
 module.exports.destroy=async function(request,response){
     //.id means converting the object id into string
 
    try{
        let post=await Post.findById(request.params.id);
 
-       //if(post.user == request.user.id){//authorization if the delete request is from the same person who created it or not
+        if(post.user == request.user.id){//authorization if the delete request is from the same person who created it or not
            post.remove();
        //deleteMany is the fxn which deleted all the comments based on some query past
            await Comment.deleteMany({post: request.params.id});//delete comments with post having that post id
            
-        //    if(request.xhr){
-        //        return response.status(200).json({
-        //            data: {
-        //                post_id : request.params.id
-        //            },
-        //            message: "Post deleted!"
-        //        });
-        //    }
-        //    request.flash('success','Post and associated comments deleted!');
-        //    return response.redirect('back');
         return response.status(200).json({
             message:"Post and associated comments deleted successfully"
         });
 
-    //    }else{
-    //        request.flash('error','You cannot delete this post');
-    //        return response.redirect('back');
-    //    }
+       }else{
+            return response.status(401).json({
+            message:"You cannot delete this post"
+            });
+       }
    }catch(err){
        //console.log('Error',err);
     //    request.flash('error',err);
